Add dynamic padding to header, refactor out App component into its component
